Fix import resolvers in eslint
