+++ .eslint-rules/imports/enforced.js
@@ -22,7 +22,7 @@ module.exports = {
       "never",
       {
         ignorePackages: true,
-        pattern: { json: "always", mp3: "always" }
+        pattern: { json: "always", mp3: "always" },
       },
     ],
   },
