+++ app/javascript/src/components/Dashboard/Settings/constants.js
@@ -1,33 +1,5 @@
 import * as yup from "yup";
 
-import Email from "./Email";
-import Password from "./Password";
-import Profile from "./Profile";
-
-export const SETTINGS_NAVLINKS = [
-  {
-    key: "profile",
-    label: "Profile",
-    description: "Manage user",
-    path: "/settings?tab=profile",
-    component: Profile,
-  },
-  {
-    key: "email",
-    label: "Email",
-    description: "Manage email",
-    path: "/settings?tab=email",
-    component: Email,
-  },
-  {
-    key: "password",
-    label: "Password",
-    description: "Manage password",
-    path: "/settings?tab=password",
-    component: Password,
-  },
-];
-
 export const PROFILE_FORM_VALIDATION_SCHEMA = yup.object().shape({
   firstName: yup.string().required("Required"),
   lastName: yup.string().required("Required"),
