+++ app/javascript/src/components/Dashboard/Settings/utils.js
@@ -1,4 +1,4 @@
-import { SETTINGS_NAVLINKS } from "./constants";
+import { SETTINGS_NAVLINKS } from "./navLinks";
 
 export const getActiveNavLink = key =>
   SETTINGS_NAVLINKS.find(navlink => key === navlink.key);
